+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* Add `PaymentMethodCustomerDataUpdated` webhook notification support
+
 ## 3.9.0 
 * Add support for passing a custom [http agent](https://nodejs.org/api/http.html#class-httpagent) in Configuration (thanks @blugavere & @James1x0!)